carrion-crow assignment
I was asked to demonstrate how to create a class constructor. The program uses two external header files. One allows the program to be restarted or exited, and the other establishes a Crow class. The species example in the code is the Carrion Crow.

The program creates a Crow object from the Crow Class.
The user is then asked to specify the color and gender of the crow. This field demonstrates input limitations (in this case, m or f is the only valid answer - the code can be modified to include other answers).
The AskToExit prompt could use some work.