demonstrates usage of linear, binary search
The program has a handful of hard-coded lottery numbers in them. Each one is five digits long and there are ten.

When the program starts, it will ask the user if they wish to a linear search (1) or binary search (2). If you type a number listed under the Winning Numbers integer array , you will get the message "Your lottery ticket is a winning ticket!". Otherwise, you'll get a message "The number you entered was not a winning ticket."

Like other programs I've uploaded, you'll be asked if you want to reload the program.